Create e2e/test_browse.py
